requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.6
django-js-asset==2.1.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1
